Outsource the api settings to getApi and create api folder.
