+++ frontend/src/api/UserAPI.js
@@ -1,20 +1,6 @@
-import axios from 'axios';
+import getApi from './getApi';
 
-const api = axios.create({
-    baseURL: `/api/`
-})
-
-api.interceptors.request.use(
-    request => {
-        const token = sessionStorage.getItem('token');
-        if (token) {
-            request.headers = {...request.headers,  Authorization: 'Bearer ' + token}
-        }
-        return request;
-    }, error => {
-        return Promise.reject(error);
-    }
-)
+let api = getApi();
 
 async function loginUser(credentials) {
     try {
